chore: hoist annotation
