+++ docs/reference/annotation.md
@@ -1,0 +1,1 @@
+::: kolena.annotation
