+++ kolena/workflow/_validators.py
@@ -25,9 +25,9 @@ from kolena._utils.datatypes import DATA_TYPE_FIELD
 from kolena._utils.datatypes import DataObject
 from kolena._utils.datatypes import get_args
 from kolena._utils.datatypes import get_origin
+from kolena.annotation import _ANNOTATION_TYPES
 from kolena.errors import WorkflowMismatchError
 from kolena.workflow._datatypes import _SCALAR_TYPES
-from kolena.workflow.annotation import _ANNOTATION_TYPES
 from kolena.workflow.asset import _ASSET_TYPES
 
 _SUPPORTED_FIELD_TYPES = [*_SCALAR_TYPES, *_ANNOTATION_TYPES, *_ASSET_TYPES, ThresholdedMetrics]
